+++ python/composio/core/models/__init__.py
@@ -1,12 +1,13 @@
 from .auth_configs import AuthConfigs
 from .connected_accounts import ConnectedAccounts
-from .custom_tool import ExperimentalAPI, ExperimentalToolkit
+from .custom_tool import ExperimentalToolkit
 from .custom_tool_types import (
     CustomTool,
     RegisteredCustomTool,
     RegisteredCustomToolkit,
     SessionContext,
 )
+from .experimental import ExperimentalAPI
 from .mcp import MCP
 from .tool_router import ToolRouter
 from .tool_router_constants import SESSION_PRESET_DIRECT_TOOLS
