+++ python/composio/core/models/custom_tool.py
@@ -479,177 +479,6 @@ class ExperimentalToolkit:
 
 
 # ────────────────────────────────────────────────────────────────
-# ExperimentalAPI — the composio.experimental namespace
-# ────────────────────────────────────────────────────────────────
-
-
-class ExperimentalAPI:
-    """Experimental APIs accessed via ``composio.experimental``.
-
-    Provides decorators for creating custom tools and toolkits that run
-    in-process alongside remote Composio tools, plus experimental SDK
-    methods whose shape may change in future releases.
-    """
-
-    Toolkit = ExperimentalToolkit
-
-    def __init__(self, client: t.Optional[t.Any] = None) -> None:
-        self._client = client
-
-    def update_acl(
-        self,
-        nanoid: str,
-        *,
-        allow_all_users: t.Optional[bool] = None,
-        allowed_user_ids: t.Optional[t.List[str]] = None,
-        not_allowed_user_ids: t.Optional[t.List[str]] = None,
-    ) -> t.Any:
-        """
-        Update the per-user ACL on a SHARED connected account. Experimental —
-        shape may change in future releases.
-
-        Only valid on SHARED connections; raises
-        ``ComposioAclOnlyForSharedError`` on a PRIVATE connection. Omit a
-        parameter to leave it unchanged; pass an empty list to clear an
-        allow/deny list. At least one parameter must be provided.
-
-        :param nanoid: The connected account ID (``ca_xxx``).
-        :param allow_all_users: When True, any ``user_id`` may use this
-            SHARED connection (subject to the deny list).
-        :param allowed_user_ids: Explicit list of allowed ``user_id`` strings.
-            Pass ``[]`` to clear.
-        :param not_allowed_user_ids: Explicit deny list (wins over allow on
-            conflict). Pass ``[]`` to clear — note that clearing the deny
-            list silently re-grants access to previously-blocked users.
-        :return: Response with ``id``, ``status``, and ``success``.
-
-        Example:
-            composio.experimental.update_acl(
-                'ca_abc',
-                allow_all_users=True,
-                not_allowed_user_ids=['user_bob'],
-            )
-        """
-        from composio_client import BadRequestError, omit
-
-        from composio import exceptions
-
-        if self._client is None:
-            raise exceptions.ValidationError(
-                "update_acl requires a Composio client. Access it via "
-                "composio.experimental.update_acl(...)."
-            )
-        if (
-            allow_all_users is None
-            and allowed_user_ids is None
-            and not_allowed_user_ids is None
-        ):
-            raise exceptions.ValidationError(
-                "update_acl requires at least one of allow_all_users, "
-                "allowed_user_ids, or not_allowed_user_ids"
-            )
-
-        acl: t.Dict[str, t.Any] = {}
-        if allow_all_users is not None:
-            acl["allow_all_users"] = allow_all_users
-        if allowed_user_ids is not None:
-            acl["allowed_user_ids"] = allowed_user_ids
-        if not_allowed_user_ids is not None:
-            acl["not_allowed_user_ids"] = not_allowed_user_ids
-
-        try:
-            return self._client.connected_accounts.patch(
-                nanoid,
-                experimental={"acl_config_for_shared": acl},
-                alias=omit,
-                connection=omit,
-            )
-        except BadRequestError as error:
-            message = str(error)
-            if "acl_config_for_shared is only valid on SHARED" in message:
-                raise exceptions.ComposioAclOnlyForSharedError(message) from error
-            raise
-
-    @t.overload
-    def tool(self, fn: t.Callable[..., t.Any], /) -> CustomTool: ...
-
-    @t.overload
-    def tool(
-        self,
-        *,
-        slug: t.Optional[str] = None,
-        name: t.Optional[str] = None,
-        description: t.Optional[str] = None,
-        extends_toolkit: t.Optional[str] = None,
-        output_params: t.Optional[t.Type[BaseModel]] = None,
-        preload: t.Optional[bool] = None,
-    ) -> t.Callable[[t.Callable[..., t.Any]], CustomTool]: ...
-
-    def tool(
-        self,
-        fn: t.Optional[t.Callable[..., t.Any]] = None,
-        *,
-        slug: t.Optional[str] = None,
-        name: t.Optional[str] = None,
-        description: t.Optional[str] = None,
-        extends_toolkit: t.Optional[str] = None,
-        output_params: t.Optional[t.Type[BaseModel]] = None,
-        preload: t.Optional[bool] = None,
-    ) -> t.Union[CustomTool, t.Callable[[t.Callable[..., t.Any]], CustomTool]]:
-        """Decorator to create a custom tool from a function.
-
-        Infers slug, name, description, and input_params from the function.
-        Override any with explicit keyword arguments.
-
-        Examples::
-
-            # Bare decorator — no parens
-            @composio.experimental.tool
-            def grep(input: GrepInput, ctx):
-                \"\"\"Search for a pattern.\"\"\"
-                return {"matches": []}
-
-            # With parens — no args
-            @composio.experimental.tool()
-            def grep(input: GrepInput, ctx):
-                \"\"\"Search for a pattern.\"\"\"
-                return {"matches": []}
-
-            # With extends_toolkit — inherits auth
-            @composio.experimental.tool(extends_toolkit="gmail")
-            def create_draft(input: DraftInput, ctx):
-                \"\"\"Create a Gmail draft.\"\"\"
-                return ctx.proxy_execute(toolkit="gmail", ...)
-        """
-
-        def decorator(f: t.Callable[..., t.Any]) -> CustomTool:
-            annotation_locals = _get_caller_locals()
-            return _infer_tool_from_function(
-                f,
-                slug=slug,
-                name=name,
-                description=description,
-                extends_toolkit=extends_toolkit,
-                output_params=output_params,
-                preload=preload,
-                annotation_locals=annotation_locals,
-            )
-
-        if fn is not None:
-            return _infer_tool_from_function(
-                fn,
-                slug=slug,
-                name=name,
-                description=description,
-                extends_toolkit=extends_toolkit,
-                output_params=output_params,
-                preload=preload,
-                annotation_locals=_get_caller_locals(),
-            )
-        return decorator
-
-
-# ────────────────────────────────────────────────────────────────
 # Serialization (for backend API payload)
 # ────────────────────────────────────────────────────────────────
